Rename global --namespace and -n flags
The --namespace and -n flags are defined in some of the subcommands and
also as a global flag. These two definitions are incompatible.
For example, when running the `sidecar-proxy egress` subcommand, it's
not possible to set the namespace where backyards is installed.

This commit deprecates the global --namespace and -n flags and
introduces the new global --backyards-namespace flag (without a
shorthand form).

Using the --namespace and -n flags as global flags is still possible
when a subcommand does not override these. In this case the following
message will be printed:

    Flag shorthand -n has been deprecated, please use --backyards-namespace instead
    Flag --namespace has been deprecated, please use --backyards-namespace instead

Also regenerated the docs and updated the `backyards --help` output in
README.md