Quick fix for renamed istio-operator resources
This is a quick fix for istio-operator#384 which would cause two
operators to run when upgrading a pre-rename istio-operator to a
post-rename operator.

In the future this should be properly fixed, so that backyards is using
the fixed istio-operator resource names, too.

Note: this requires a not yet ready fix in istio-operator, too